TMS-2: Add tests for invalid zip code on Sign Up page
