--- src/main/java/example/org/SignUpTest.java
@@ -1,27 +1,0 @@
-package example.org;
-
-
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.Test;
-
-import static org.testng.Assert.assertTrue;
-
-public class SignUpTest {
-
-    @Test
-    public void zipCodeShouldHave5Digits() {
-        System.setProperty("webdriver.chrome.driver", "src/main/resources/chromedriver.exe");
-        WebDriver driver = new ChromeDriver();
-        driver.get("https://www.sharelane.com/cgi-bin/register.py");
-        WebElement zipCodeField = driver.findElement(By.name("zip_code"));
-        zipCodeField.sendKeys("12345");
-        driver.findElement(By.cssSelector("input[value='Continue']")).click();
-        WebElement registerButton = driver.findElement(By.cssSelector("input[value='Register']"));
-        assertTrue(registerButton.isDisplayed(), "User was not redirected to the registration page");
-        driver.quit();
-    }
-}
